#5 feat: refactored vacancies service
