--- backend/src/Services/Vacancies/Vacancies.Application/Drivers/Services/IActivateDriver.cs
@@ -1,9 +1,0 @@
-﻿using OpenQA.Selenium;
-
-namespace Vacancies.Application.Drivers.Services
-{
-    public interface IActivateDriver
-    {
-        Task<IWebDriver> ActivateScrapingDriver();
-    }
-}
