--- backend/src/Services/Vacancies/Vacancies.Application/Vacancies/Dto/VacancyResponseModel.cs
@@ -1,11 +1,0 @@
-﻿namespace Vacancies.Application.Vacancies.Dto
-{
-    public class VacancyResponseModel
-    {
-        public Guid Gid { get; set; }
-        public string Name { get; set; }
-        public string Description { get; set; }
-        public string Location { get; set; }
-        public string? Salary { get; set; }
-    }
-}
