+++ backend/src/Services/Vacancies/Vacancies.Core/Consts/VacanciesConsts.cs
@@ -1,0 +1,6 @@
+﻿namespace Vacancies.Core.Consts;
+
+public static class VacanciesConsts
+{
+    public static string DjinniUrl => "https://djinni.co/";
+}
