+++ backend/src/Services/Vacancies/Vacancies.Core/Entities/SocialRequest.cs
@@ -1,14 +1,23 @@
 ﻿using Shared.Abstractions.Primitives;
 
-namespace Vacancies.Core.Entities
+namespace Vacancies.Core.Entities;
+
+public class SocialRequest : Entity
 {
-    public class SocialRequest : Entity
+    private readonly List<Vacancy> _vacancies = new();
+
+    private SocialRequest(string uri) : base()
     {
-        private readonly List<Vacancy> _vacancies = new();
-        public SocialRequest() : base() { }
+        Uri = uri;
+    }
 
-        public string Uri { get; set; }
+    public string Uri { get; private set; }
 
-        public ICollection<Vacancy>? Vacancies => _vacancies;
+    public ICollection<Vacancy>? Vacancies => _vacancies;
+
+    public static SocialRequest Init(
+        string uri)
+    {
+        return new SocialRequest(uri);
     }
 }
