--- backend/src/Services/Vacancies/Vacancies.Core/Enums/DjinniConst.cs
@@ -1,13 +1,0 @@
-﻿namespace Vacancies.Core.Enums
-{
-    public static class DjinniConst
-    {
-        public const string DjinniUrl = "https://djinni.co/";
-        public enum Keywords
-        {
-            primary_keyword,
-            exp_level,
-            region
-        }
-    }
-}
