+++ backend/src/Services/Vacancies/Vacancies.Core/Enums/Keywords.cs
@@ -1,0 +1,8 @@
+﻿namespace Vacancies.Core.Enums;
+
+public enum Keywords
+{
+    primary_keyword,
+    exp_level,
+    region
+}
