+++ backend/src/Services/Vacancies/Vacancies.Core/Responses/VacancyResponse.cs
@@ -1,0 +1,3 @@
+﻿namespace Vacancies.Core.Responses;
+
+public record VacancyResponse(Guid Gid, string Name, string Description, string Location, string Salary);
