+++ backend/src/Services/Vacancies/Vacancies.Core/Vacancies.Core.csproj
@@ -7,7 +7,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+	<PackageReference Include="HtmlAgilityPack" Version="1.11.46" />
     <ProjectReference Include="..\..\..\SharedKernel\Shared.Abstractions\Shared.Abstractions.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Selenium.WebDriver" Version="4.8.1" />
   </ItemGroup>
 
 </Project>
