+++ backend/src/Services/Vacancies/Vacancies.Persistence/Configurations/SocialRequestConfiguration.cs
@@ -2,14 +2,13 @@
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Vacancies.Core.Entities;
 
-namespace Vacancies.Persistence.Configurations
+namespace Vacancies.Persistence.Configurations;
+
+internal sealed class SocialRequestConfiguration : IEntityTypeConfiguration<SocialRequest>
 {
-    public class SocialRequestConfiguration : IEntityTypeConfiguration<SocialRequest>
+    public void Configure(EntityTypeBuilder<SocialRequest> builder)
     {
-        public void Configure(EntityTypeBuilder<SocialRequest> builder)
-        {
-            builder.HasKey(x => x.Gid);
-            builder.Property(x => x.Uri).IsRequired();
-        }
+        builder.HasKey(x => x.Gid);
+        builder.Property(x => x.Uri).IsRequired();
     }
 }
