+++ backend/src/Services/Vacancies/Vacancies.Persistence/VacanciesDbContext.cs
@@ -1,16 +1,15 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Vacancies.Core.Entities;
 
-namespace Vacancies.Persistence
+namespace Vacancies.Persistence;
+
+public sealed class VacanciesDbContext : DbContext
 {
-    public class VacanciesDbContext : DbContext
-    {
-        public DbSet<Vacancy>? Vacancies { get; set; }
-        public DbSet<SocialRequest>? SocialRequests { get; set; }
+    public DbSet<Vacancy>? Vacancies { get; set; }
+    public DbSet<SocialRequest>? SocialRequests { get; set; }
 
-        public VacanciesDbContext(DbContextOptions<VacanciesDbContext> options) : base(options) { }
+    public VacanciesDbContext(DbContextOptions<VacanciesDbContext> options) : base(options) { }
 
-        protected override void OnModelCreating(ModelBuilder builder)
-            => builder.ApplyConfigurationsFromAssembly(GetType().Assembly);
-    }
+    protected override void OnModelCreating(ModelBuilder builder)
+        => builder.ApplyConfigurationsFromAssembly(GetType().Assembly);
 }
